+++ apps/client/src/features/app-settings/panel/project-panel/ProjectPanel.tsx
@@ -1,40 +1,15 @@
 import { useState } from 'react';
 import { Button } from '@chakra-ui/react';
 
-import { PROJECT_LIST } from '../../../../common/api/apiConstants';
-import { maybeAxiosError } from '../../../../common/api/apiUtils';
-import { createProject } from '../../../../common/api/ontimeApi';
-import { ontimeQueryClient } from '../../../../common/queryClient';
 import * as Panel from '../PanelUtils';
 
-import ProjectForm, { ProjectFormValues } from './ProjectForm';
 import ProjectList from './ProjectList';
-
-import style from './ProjectPanel.module.scss';
 
 export default function ProjectPanel() {
   const [isCreatingProject, setIsCreatingProject] = useState(false);
-  const [submitError, setSubmitError] = useState<string | null>(null);
 
   const handleToggleCreate = () => {
     setIsCreatingProject((prev) => !prev);
-    setSubmitError(null);
-  };
-
-  const handleSubmitCreate = async (values: ProjectFormValues) => {
-    try {
-      setSubmitError(null);
-      const filename = values.filename.trim();
-      if (!filename) {
-        setSubmitError('Project name cannot be empty');
-        return;
-      }
-      await createProject(filename);
-      await ontimeQueryClient.invalidateQueries({ queryKey: PROJECT_LIST });
-      handleToggleCreate();
-    } catch (error) {
-      maybeAxiosError(error);
-    }
   };
 
   return (
@@ -43,30 +18,12 @@ export default function ProjectPanel() {
       <Panel.Section>
         <Panel.Card>
           <Panel.SubHeader>
-            {!isCreatingProject ? (
-              <>
-                Manage projects
-                <Button variant='ontime-filled' onClick={handleToggleCreate}>
-                  New
-                </Button>
-              </>
-            ) : (
-              <span>Create new project</span>
-            )}
+            Manage projects
+            <Button variant='ontime-filled' onClick={handleToggleCreate}>
+              New
+            </Button>
           </Panel.SubHeader>
-          {isCreatingProject ? (
-            <div className={style.createContainer}>
-              <ProjectForm
-                action='create'
-                filename=''
-                onSubmit={handleSubmitCreate}
-                onCancel={handleToggleCreate}
-                submitError=''
-              />
-              {submitError && <span className={style.createSubmitError}>{submitError}</span>}
-            </div>
-          ) : null}
-          {!isCreatingProject && <ProjectList />}
+          <ProjectList onToggleCreate={handleToggleCreate} isCreatingProject={isCreatingProject} />
         </Panel.Card>
       </Panel.Section>
     </>
